+++ scripts/travis/fuzz.sh
@@ -3,7 +3,7 @@
 ROOT_DIR=$(dirname $0)/../../
 INCLUDE_DIR=${ROOT_DIR}/src/
 FUZZING_DIR=${ROOT_DIR}/fuzzing/
-CXXFLAGS="-g -fprofile-instr-generate -fcoverage-mapping -fsanitize=address,fuzzer"
+CXXFLAGS="-g -fprofile-instr-generate -fcoverage-mapping -fsanitize=address,undefined,fuzzer -fno-sanitize-recover=all"
 
 fuzz() {
 	NAME="$1"