+++ src/ArduinoJson/Numbers/parseInteger.hpp
@@ -4,34 +4,16 @@
 
 #pragma once
 
-#include "../Configuration.hpp"
-#include "../Polyfills/ctype.hpp"
+#include "../Polyfills/type_traits.hpp"
+#include "convertNumber.hpp"
+#include "parseNumber.hpp"
 
 namespace ARDUINOJSON_NAMESPACE {
 template <typename T>
 T parseInteger(const char *s) {
-  if (!s) return 0;  // NULL
-
-  if (*s == 't') return 1;  // "true"
-
-  T result = 0;
-  bool negative_result = false;
-
-  switch (*s) {
-    case '-':
-      negative_result = true;
-      s++;
-      break;
-    case '+':
-      s++;
-      break;
-  }
-
-  while (isdigit(*s)) {
-    result = T(result * 10 + T(*s - '0'));
-    s++;
-  }
-
-  return negative_result ? T(~result + 1) : result;
+  // try to reuse the same parameters as JsonDeserializer
+  typedef typename choose_largest<UInt, typename make_unsigned<T>::type>::type
+      TUInt;
+  return parseNumber<Float, TUInt>(s).template as<T>();
 }
 }  // namespace ARDUINOJSON_NAMESPACE