+++ src/ArduinoJson/Polyfills/attributes.hpp
@@ -33,3 +33,13 @@
 #else
 #define NOEXCEPT throw()
 #endif
+
+#if defined(__has_attribute)
+#if __has_attribute(no_sanitize)
+#define ARDUINOJSON_NO_SANITIZE(check) __attribute__((no_sanitize(check)))
+#else
+#define ARDUINOJSON_NO_SANITIZE(check)
+#endif
+#else
+#define ARDUINOJSON_NO_SANITIZE(check)
+#endif