+++ src/ArduinoJson/Polyfills/type_traits.hpp
@@ -15,5 +15,6 @@
 #include "type_traits/is_same.hpp"
 #include "type_traits/is_signed.hpp"
 #include "type_traits/is_unsigned.hpp"
+#include "type_traits/make_unsigned.hpp"
 #include "type_traits/remove_const.hpp"
 #include "type_traits/remove_reference.hpp"