+++ test/JsonDeserializer/misc.cpp
@@ -7,11 +7,6 @@
 
 using namespace Catch::Matchers;
 
-namespace my {
-using ARDUINOJSON_NAMESPACE::isinf;
-using ARDUINOJSON_NAMESPACE::isnan;
-}  // namespace my
-
 TEST_CASE("deserializeJson(DynamicJsonDocument&)") {
   DynamicJsonDocument doc(4096);
 
@@ -45,64 +40,6 @@ TEST_CASE("deserializeJson(DynamicJsonDocument&)") {
       DeserializationError err = deserializeJson(doc, "%*$£¤");
 
       REQUIRE(err == DeserializationError::InvalidInput);
-    }
-  }
-
-  SECTION("Integers") {
-    SECTION("0") {
-      DeserializationError err = deserializeJson(doc, "0");
-      REQUIRE(err == DeserializationError::Ok);
-      REQUIRE(doc.is<int>() == true);
-      REQUIRE(doc.as<int>() == 0);
-      REQUIRE(doc.as<std::string>() == "0");  // issue #808
-    }
-
-    SECTION("Negative") {
-      DeserializationError err = deserializeJson(doc, "-42");
-
-      REQUIRE(err == DeserializationError::Ok);
-      REQUIRE(doc.is<int>());
-      REQUIRE_FALSE(doc.is<bool>());
-      REQUIRE(doc.as<int>() == -42);
-    }
-  }
-
-  SECTION("Floats") {
-    SECTION("Double") {
-      DeserializationError err = deserializeJson(doc, "-1.23e+4");
-
-      REQUIRE(err == DeserializationError::Ok);
-      REQUIRE_FALSE(doc.is<int>());
-      REQUIRE(doc.is<double>());
-      REQUIRE(doc.as<double>() == Approx(-1.23e+4));
-    }
-
-    SECTION("NaN") {
-      DeserializationError err = deserializeJson(doc, "NaN");
-      REQUIRE(err == DeserializationError::Ok);
-      REQUIRE(doc.is<float>() == true);
-      REQUIRE(my::isnan(doc.as<float>()));
-    }
-
-    SECTION("Infinity") {
-      DeserializationError err = deserializeJson(doc, "Infinity");
-      REQUIRE(err == DeserializationError::Ok);
-      REQUIRE(doc.is<float>() == true);
-      REQUIRE(my::isinf(doc.as<float>()));
-    }
-
-    SECTION("+Infinity") {
-      DeserializationError err = deserializeJson(doc, "+Infinity");
-      REQUIRE(err == DeserializationError::Ok);
-      REQUIRE(doc.is<float>() == true);
-      REQUIRE(my::isinf(doc.as<float>()));
-    }
-
-    SECTION("-Infinity") {
-      DeserializationError err = deserializeJson(doc, "-Infinity");
-      REQUIRE(err == DeserializationError::Ok);
-      REQUIRE(doc.is<float>() == true);
-      REQUIRE(my::isinf(doc.as<float>()));
     }
   }