+++ test/Numbers/parseFloat.cpp
@@ -2,6 +2,8 @@
 // Copyright Benoit Blanchon 2014-2019
 // MIT License
 
+#define ARDUINOJSON_USE_DOUBLE 0
+
 #include <ArduinoJson/Numbers/parseFloat.hpp>
 #include <catch.hpp>
 
@@ -33,10 +35,6 @@ void checkInf(const char* input, bool negative) {
 }
 
 TEST_CASE("parseFloat<float>()") {
-  SECTION("Null") {
-    check<float>(NULL, 0);
-  }
-
   SECTION("Float_Short_NoExponent") {
     check<float>("3.14", 3.14f);
     check<float>("-3.14", -3.14f);
@@ -97,19 +95,13 @@ TEST_CASE("parseFloat<float>()") {
     checkInf<float>("inf", false);
     checkInf<float>("+inf", false);
     checkInf<float>("-inf", true);
-  }
 
-  SECTION("Boolean") {
-    check<float>("false", 0.0f);
-    check<float>("true", 1.0f);
+    checkInf<float>("1e300", false);
+    checkInf<float>("-1e300", true);
   }
 }
 
 TEST_CASE("parseFloat<double>()") {
-  SECTION("Null") {
-    check<double>(NULL, 0);
-  }
-
   SECTION("Short_NoExponent") {
     check<double>("3.14", 3.14);
     check<double>("-3.14", -3.14);
@@ -168,10 +160,5 @@ TEST_CASE("parseFloat<double>()") {
   SECTION("NaN") {
     checkNaN<double>("NaN");
     checkNaN<double>("nan");
-  }
-
-  SECTION("Boolean") {
-    check<double>("false", 0.0);
-    check<double>("true", 1.0);
   }
 }