+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ArduinoJson: change log
 =======================
 
+HEAD
+----
+
+* Fixed an integer overflow in the JSON deserializer
+* Added overflow handling in `JsonVariant::as<T>()` and `JsonVariant::is<T>()`.
+   - `as<T>()` returns `0` if the integer `T` overflows
+   - `is<T>()` returns `false` if the integer `T` overflows
+
 v6.9.1 (2019-03-01)
 ------